Override the notifications protocol with environment-specific URL.
Although clients should follow the next url from the notifications feed,
for PAM this needs to be modified according to the environment config,
as we want to hit specific clusters rather than load balancers, and as
we are really checking the publish path, we prefer to bypass Varnish and
the API policy component, so use the `__servicename` path prefix in the
configuration. In this context, "follow the next url" really means "use
the whole query string".