:bug: Prevent deadlock; initialise file-based config immediately.
- When using the file-based config, validator credentials were never being read because of a deadlock in the use of ThreadSafeEnvironments.
- Read file-based config immediately rather than waiting for 1 minute, by simulating a timer with zero initial delay.

For both etcd and file-based configs, I think there remains a theoretical race condition, as each is initialised by a goroutine that is not guaranteed to execute prior to the configuration of healthchecks. Th healthchecks _should_ be modified dynamically whenever the environments change, but aren't. However, environments change rarely; so long as credential changes are picked up and applied, this flaw can remain, as it's complex to fix and we may switch to a different monitoring solution anyway.